fix(tests): move integration tests inside crate
