+++ app/src/integration_tests/eth_sign.rs
@@ -1,5 +1,5 @@
 /*******************************************************************************
-*   (c) 2021 Zondax GmbH
+*   (c) 2022 Zondax AG
 *
 *  Licensed under the Apache License, Version 2.0 (the "License");
 *  you may not use this file except in compliance with the License.
@@ -13,12 +13,7 @@
 *  See the License for the specific language governing permissions and
 *  limitations under the License.
 ********************************************************************************/
-mod prelude;
-
-
-
-
-
+use super::prelude::*;
 
 #[cfg(feature = "blind-sign")]
 use constants::INS_ETH_BLIND_SIGN as INS;
