+++ app/src/integration_tests/wallet_id.rs
@@ -1,5 +1,5 @@
 /*******************************************************************************
-*   (c) 2021 Zondax GmbH
+*   (c) 2022 Zondax AG
 *
 *  Licensed under the Apache License, Version 2.0 (the "License");
 *  you may not use this file except in compliance with the License.
@@ -13,8 +13,7 @@
 *  See the License for the specific language governing permissions and
 *  limitations under the License.
 ********************************************************************************/
-mod prelude;
-use prelude::*;
+use super::prelude::*;
 
 use constants::{INS_GET_WALLET_ID as INS, WALLET_ID_LEN};
